requirements.txt
attrs==18.2.0
Automat==0.7.0
certifi==2018.10.15
chardet==3.0.4
constantly==15.1.0
Django==2.1.2
django-pure-pagination==0.3.0
docopt==0.6.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
mysqlclient==1.3.13
Pillow==5.3.0
pipreqs==0.4.9
PyHamcrest==1.9.0
pytz==2018.5
requests==2.19.1
six==1.11.0
Twisted==18.7.0
urllib3==1.23
yarg==0.1.9
zope.interface==4.5.0
